Refactor StandardKeyBindingResponding tests
